first touch with react and api endpoints. more files to clean visualization.
